+++ lib/crono/config.rb
@@ -7,12 +7,17 @@ module Crono
 
     attr_accessor :cronotab,
                   :environment,
+                  :verbose,
                   :lifecycle_events
 
     def initialize
       self.cronotab = CRONOTAB
       self.environment = ENV['RAILS_ENV'] || ENV['RACK_ENV'] || 'development'
-      self.lifecycle_events = { startup: [] }
+      self.verbose = false
+      self.lifecycle_events = {
+        startup: [],
+        shutdown: [],
+      }
     end
 
     def on(event, &block)